Reorganize functions and fix formatting
